Add documentation & break out code into modules
