+++ src/options.rs
@@ -3,6 +3,7 @@ use std::ffi::CString;
 
 type LedMatrixOptionsResult = Result<(), &'static str>;
 
+/// The Rust representation of LedMatrixOptions, which contains parameters to specify your hardware setup.
 #[derive(Debug)]
 #[repr(C)]
 pub struct LedMatrixOptions {
@@ -27,6 +28,7 @@ pub struct LedMatrixOptions {
     pub(crate) limit_refresh_rate_hz: c_int,
 }
 
+/// The Rust representation of LedRuntimeOptions, which contains parameters to specify how the library behaves at runtime.
 #[derive(Debug)]
 #[repr(C)]
 pub struct LedRuntimeOptions {
@@ -37,6 +39,14 @@ pub struct LedRuntimeOptions {
 }
 
 impl LedMatrixOptions {
+    /// Creates a new `LedMatrixOptions` struct with the default parameters.
+    ///
+    /// ```
+    /// use rpi_led_matrix::{LedMatrix,LedMatrixOptions};
+    /// let mut options = LedMatrixOptions::new();
+    /// options.set_hardware_mapping("adafruit-hat-pwm");
+    /// let matrix = LedMatrix::new(Some(options), None).unwrap();
+    /// ```
     pub fn new() -> LedMatrixOptions {
         LedMatrixOptions {
             hardware_mapping: CString::new("regular").unwrap().into_raw(),
@@ -61,6 +71,7 @@ impl LedMatrixOptions {
         }
     }
 
+    /// Sets the type of GPIO mapping used (e.g., "adafruit-hat-pwm").
     pub fn set_hardware_mapping(&mut self, mapping: &str) {
         unsafe {
             let _ = CString::from_raw(self.hardware_mapping);
@@ -68,22 +79,27 @@ impl LedMatrixOptions {
         }
     }
 
+    /// Sets the number of rows on the panels being used. Typically 8, 16, 32 or 64.
     pub fn set_rows(&mut self, rows: u32) {
         self.rows = rows as c_int;
     }
 
+    /// Sets the number of columns on the panels being used. Typically 32 or 64.
     pub fn set_cols(&mut self, cols: u32) {
         self.cols = cols as c_int;
     }
 
+    /// Sets the number of panels daisy-chained together.
     pub fn set_chain_length(&mut self, chain_length: u32) {
         self.chain_length = chain_length as c_int;
     }
 
+    /// Sets the number of parallel chains. Valid range: [1,3].
     pub fn set_parallel(&mut self, parallel: u32) {
         self.parallel = parallel as c_int;
     }
 
+    /// Sets the number of PWM bits to use. Valid range: [1,11].
     pub fn set_pwm_bits(&mut self, pwm_bits: u8) -> LedMatrixOptionsResult {
         if pwm_bits > 11 {
             Err("Pwm bits can only have value between 0 and 11 inclusive")
@@ -93,10 +109,12 @@ impl LedMatrixOptions {
         }
     }
 
+    /// Sets the number of nanoseconds of delay for our LSB
     pub fn set_pwm_lsb_nanoseconds(&mut self, pwm_lsb_nanoseconds: u32) {
         self.pwm_lsb_nanoseconds = pwm_lsb_nanoseconds as c_int;
     }
 
+    /// Sets the pannel brightness in percent.
     pub fn set_brightness(&mut self, brightness: u8) -> LedMatrixOptionsResult {
         if brightness > 100 || brightness < 1 {
             Err("Brigthness can only have value between 1 and 100 inclusive")
@@ -106,10 +124,12 @@ impl LedMatrixOptions {
         }
     }
 
+    /// Sets the scan mode. 0: progressive, 1: interlaced.
     pub fn set_scan_mode(&mut self, scan_mode: u32) {
         self.scan_mode = scan_mode as c_int;
     }
 
+    /// Sets the ordering of the LEDs on your panel.
     pub fn set_led_rgb_sequence(&mut self, sequence: &str) {
         unsafe {
             let _ = CString::from_raw(self.led_rgb_sequence);
@@ -117,6 +137,14 @@ impl LedMatrixOptions {
         }
     }
 
+    /// Semicolon-separated list of pixel-mappers to arrange pixels (e.g. "U-mapper;Rotate:90").
+    ///
+    /// Valid mapping options
+    ///
+    /// * Mirror
+    /// * Rotate:<Angle>
+    /// * U-mapper
+    /// * V-mapper
     pub fn set_pixel_mapper_config(&mut self, mapper: &str) {
         unsafe {
             let _ = CString::from_raw(self.pixel_mapper_config);
@@ -124,6 +152,7 @@ impl LedMatrixOptions {
         }
     }
 
+    /// Sets if hardware pin-pulse generation should be used.
     pub fn set_hardware_pulsing(&mut self, enable: bool) {
         if enable {
             self.disable_hardware_pulsing = 0;
@@ -132,6 +161,7 @@ impl LedMatrixOptions {
         }
     }
 
+    /// Configures if the current refresh rate should be printed by the C++ library.
     pub fn set_refresh_rate(&mut self, enable: bool) {
         if enable {
             self.show_refresh_rate = 1;
@@ -140,6 +170,7 @@ impl LedMatrixOptions {
         }
     }
 
+    /// If set, invert the color displayed.
     pub fn set_inverse_colors(&mut self, enable: bool) {
         if enable {
             self.inverse_colors = 1;
@@ -148,22 +179,53 @@ impl LedMatrixOptions {
         }
     }
 
+    /// Sets the type of multiplexing used.
+    ///
+    /// 0.  direct
+    /// 1.  Stripe
+    /// 2.  Checkered
+    /// 3.  Spiral
+    /// 4.  ZStripe
+    /// 5.  ZnMirrorZStripe
+    /// 6.  coreman
+    /// 7.  Kaler2Scan
+    /// 8.  ZStripeUneven
+    /// 9.  P10-128x4-Z
+    /// 10. QiangLiQ8
+    /// 11. InversedZStripe
+    /// 12. P10Outdoor1R1G1-1
+    /// 13. P10Outdoor1R1G1-2
+    /// 14. P10Outdoor1R1G1-3
+    /// 15. P10CoremanMapper
+    /// 16. P8Outdoor1R1G1
     pub fn set_multiplexing(&mut self, multiplexing: u32) {
         self.multiplexing = multiplexing as c_int;
     }
 
+    /// Sets the type of row addressing to be used.
+    ///
+    /// 0. default
+    /// 1. AB-addressed panels
+    /// 2. direct row select
+    /// 3. ABC-addressed panels
+    /// 4. ABC Shift + DE direct
     pub fn set_row_addr_type(&mut self, row_addr_type: u32) {
         self.row_address_type = row_addr_type as c_int;
     }
 
+    /// Limit refresh rate to this frequency in Hz. (0 = no limit)
+    ///
+    /// Useful to keep a constant refresh rate on loaded system.
     pub fn set_limit_refresh(&mut self, limit_refresh: u32) {
         self.limit_refresh_rate_hz = limit_refresh as c_int;
     }
 
+    /// Configures how many bits to use for time-based dithering.
     pub fn set_pwm_dither_bits(&mut self, pwm_dither_bits: u32) {
         self.pwm_dither_bits = pwm_dither_bits as c_int;
     }
 
+    /// Needed to initialize special panels. Supported: 'FM6126A', 'FM6127'
     pub fn set_panel_type(&mut self, panel_type: &str) {
         unsafe {
             let _ = CString::from_raw(self.panel_type);
@@ -189,6 +251,14 @@ impl Drop for LedMatrixOptions {
 }
 
 impl LedRuntimeOptions {
+    /// Creates a new `LedRuntimeOptions` struct with the default parameters.
+    ///
+    /// ```
+    /// use rpi_led_matrix::{LedMatrix, LedRuntimeOptions};
+    /// let mut rt_options = LedRuntimeOptions::new();
+    /// rt_options.set_gpio_slowdown(3);
+    /// let matrix = LedMatrix::new(None, Some(rt_options)).unwrap();
+    /// ```
     pub fn new() -> Self {
         Self {
             gpio_slowdown: 1,
@@ -198,18 +268,22 @@ impl LedRuntimeOptions {
         }
     }
 
+    /// Sets the GPIO slowdown, in . Needed for faster Pis/slower panels
     pub fn set_gpio_slowdown(&mut self, gpio_slowdown: u32) {
         self.gpio_slowdown = gpio_slowdown as i32;
     }
 
+    /// If True, make the process run in the background as daemon.
     pub fn set_daemon(&mut self, daemon: bool) {
         self.daemon = if daemon { 1 } else { 0 };
     }
 
+    /// If True, drop privileges from 'root' after initializing the hardware.
     pub fn set_drop_privileges(&mut self, drop_privileges: bool) {
         self.drop_privileges = if drop_privileges { 1 } else { 0 };
     }
 
+    /// You almost definitely want this to be left as True. Use this if you know what you're doing.
     pub fn set_do_gpio_init(&mut self, do_gpio_init: bool) {
         self.do_gpio_init = do_gpio_init;
     }
